requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-bdd==7.0.0
pytest-html==4.1.1
python-dotenv==1.0.0
requests==2.31.0
fake-useragent==1.4.0
allure-pytest==2.14.2
allure-python-commons==2.14.2